+++ lib/cassie/support.rb
@@ -1,6 +1,6 @@
 module Cassie
   module Support
-    require_relative 'support/command_runner'
+    require_relative 'support/system_command'
     require_relative 'support/statement_parser'
   end
 end