+++ lib/cassie/support/system_command.rb
@@ -1,6 +1,6 @@
 module Cassie
   module Support
-    class CommandRunner
+    class SystemCommand
       attr_reader :binary, :args, :command, :status, :duration, :output
 
       # When a block is given, the command runs before yielding
@@ -17,6 +17,7 @@ module Cassie
       end
 
       # Runs the command
+      # @return [Boolean] true if execution completed without crashing
       def run
         t1=Time.now
 
@@ -29,36 +30,38 @@ module Cassie
         completed?
       end
 
-      # Runs the command, raising if program
-      # was not run successfully
-      def run!
-        unless run && success?
-          raise RuntimeError.new(failure_message)
-        end
+      # Runs the command, expecting an exit status of 0
+      # @return [Boolean] true if execution completed without crashing
+      # @raise [RuntimeError] if program was not run successfully
+      def succeed
+        fail unless run && success?
+
         true
       end
 
-      # Returns false if the command hasn't been executed yet
+      # @return [Boolean] false if the command hasn't been run yet
       def run?
         !!@duration
       end
 
-      # Returns the exit code for the command. Runs the command if it hasn't run yet.
+      # Runs the command if it hasn't been run yet.
+      # @return [Fixnum] the exit code for the command.
       def exitcode
         status.exitstatus
       end
 
-      # Returns true if exited 0
+      # @return [Boolean] true if command has been run, and exited with status of 0,
+      #                   otherwise returns false.
       def success?
         return false unless run?
         exitcode == 0
       end
 
-      # Returns true if the command completed execution and success bits were set.
-      # Will return false if the command hasn't been executed
+      # @return [Boolean] true if the command completed execution and success bits were set, regardless of the exit status code.
+      #                   false if the command hasn't been executed, failed to exit, or crashed.
       def completed?
         return false unless run?
-        status.exited? && @status.success?
+        status.exited? && @status.success? #status.success is NOT the exit code == 0!
       end
 
       def failure_message
@@ -71,6 +74,12 @@ module Cassie
       end
 
       protected
+
+      # raise a Runtime error with a failure message
+      # specific to the command
+      def fail
+        raise RuntimeError.new(failure_message)
+      end
 
       def red(message)
         "\e[1m\e[31m#{message}\e[0m\e[22m"